requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
twilio
kivy
requests
werkzeug
pytesseract
pillow
gunicorn
